backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.1
openai==1.45.0
httpx==0.27.2
chromadb==0.5.5
pypdf==4.3.1
